+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 - issue #685 fixed by removing all URL query params.
 
+### Patches
+
+- updated version of `solid-common-vocab` to pull in common RDF/JS types.
+
 ## 1.1.0 - 2020-11-27
 
 ### New features